fix: notification

+++ backend/src/main/java/com/GDU/backend/services/Impl/JwtService.java
@@ -19,7 +19,7 @@ public class JwtService {
     private String secretKey;
 
     @Value("${spring.application.security.jwt.expiration}")
-    private int jwtExpiration;
+    private long jwtExpiration;
 
     public String extractUsername(String token) {
         return extractClaim(token, Claims::getSubject);
